🚚 Rename: rename folders and paths
폴더이름 kebab-case로 수정, import 방식 수정
