+++ src/app/my-page/page.tsx
@@ -1,4 +1,4 @@
-import EditProfileForm from '@/widgets/editProfile/ui/EditProfileForm';
+import EditProfileForm from '@/widgets/edit-profile/ui/EditProfileForm';
 
 export default function MyProfilePage() {
   return <EditProfileForm />;
