+++ src/shared/index.tsx
@@ -1,6 +1,8 @@
 export { default as Logo } from './ui/Logo';
 export { default as Input } from './ui/Input';
 export { default as Button } from './ui/Button';
+export { default as EditableField } from './ui/EditableField';
+export { default as ImageUploader } from './ui/ImageUploader';
 export { default as useInfiniteScroll } from './hooks/useInfiniteScroll';
 export { default as useCarousel } from './hooks/useCarousel';
 export { default as calculateStayDuration } from './utils/calculateStayDuration';
