+++ src/shared/ui/Button.tsx
@@ -1,7 +1,5 @@
 'use client';
 
-import React from 'react';
-
 interface ButtonProps {
   text: string;
   onClick?: () => void;
