+++ src/shared/ui/EditableField.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import React, { useState } from 'react';
+import { useState } from 'react';
 
 interface EditableFieldProps {
   label?: string;
