+++ src/shared/ui/ImageUploader.tsx
@@ -1,6 +1,3 @@
-'use client';
-
-import React from 'react';
 import Image from 'next/image';
 
 interface ImageUploaderProps {
