+++ src/shared/ui/Input.tsx
@@ -1,5 +1,3 @@
-import React from 'react';
-
 interface InputProps {
   placeholder: string;
   type?: string;
