+++ src/widgets/travel-info/components/DetailSection.tsx
