+++ src/widgets/travel-info/components/Map.tsx
@@ -1,5 +1,3 @@
-import React from 'react';
-
 export default function Map() {
   return (
     <div className="flex h-[calc(100vh-5rem)] max-w-[calc(512px-2rem)] items-center justify-center bg-gray-300">
